+++ components/st-page-title/index.ts
@@ -1,0 +1,1 @@
+export * from './st-page-title.component';