+++ web/app/+examples/index.ts
@@ -18,3 +18,4 @@ export * from './colors';
 export * from './typography';
 export * from './icons';
 export * from './logo';
+export * from './page-title';