+++ web/app/+examples/list/index.ts
@@ -1,1 +1,1 @@
-export * from './list.component.ts';
+export * from './list.component';