+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add postCSS support
 * Update to Angular 2.1.0 and Typescript 2.0.3 with types in npm
 * Add radio menu component
+* New component Title Page
 
 ## 0.4.0 (November 02, 2016)